+++ policy_sentry/shared/config.py
@@ -4,10 +4,11 @@ Functions to support creating the proper resources in the Policy Sentry Config d
 import os
 import sys
 import shutil
+from distutils.dir_util import copy_tree
 from policy_sentry.shared.file import create_directory_if_it_doesnt_exist, \
     list_files_in_directory, read_yaml_file
 from policy_sentry.shared.constants import HOME, CONFIG_DIRECTORY, DATABASE_FILE_NAME, AUDIT_DIRECTORY_FOLDER, \
-    AUDIT_DIRECTORY_PATH
+    AUDIT_DIRECTORY_PATH, DATABASE_FILE_PATH, HTML_DIRECTORY_PATH, HTML_DATA_DIRECTORY_SUBFOLDER
 
 
 def create_policy_sentry_config_directory():
@@ -17,26 +18,47 @@ def create_policy_sentry_config_directory():
     """
     print("Creating the database...")
 
-    database_path = HOME + CONFIG_DIRECTORY + DATABASE_FILE_NAME
-    print("We will store the new database here: " + database_path)
+    print("We will store the new database here: " + DATABASE_FILE_PATH)
     # If the database file already exists
-    if os.path.exists(database_path):
-        os.remove(database_path)
+    if os.path.exists(DATABASE_FILE_PATH):
+        os.remove(DATABASE_FILE_PATH)
     elif os.path.exists(HOME + CONFIG_DIRECTORY):
         pass
     # If the config directory does not exist
     else:
         os.mkdir(HOME + CONFIG_DIRECTORY)
-    return database_path
+    return DATABASE_FILE_PATH
+
+
+def create_html_docs_directory():
+    """
+    Copies the HTML files from the pip package over to its own folder in the CONFIG_DIRECTORY.
+    Also copies over the links.yml file, which is a mapping of services and relevant HTML links in the AWS docs.
+    Essentially:
+    mkdir -p ~/.policy_sentry/data/docs
+    cp -r $MODULE_DIR/policy_sentry/shared/data/docs ~/.policy_sentry/data/docs
+    :return:
+    """
+    create_directory_if_it_doesnt_exist(HTML_DIRECTORY_PATH)
+    # Copy from the existing html docs folder - the path ./policy_sentry/shared/data/docs within this repository
+    existing_html_docs_folder = os.path.abspath(os.path.dirname(__file__)) + HTML_DATA_DIRECTORY_SUBFOLDER
+    copy_tree(existing_html_docs_folder, HTML_DIRECTORY_PATH)
+    # Copy the links.yml file from here to the config directory
+    existing_links_file = os.path.abspath(os.path.dirname(__file__)) + '/' + 'links.yml'
+    target_links_file = HOME + CONFIG_DIRECTORY + 'links.yml'
+    shutil.copy(existing_links_file, target_links_file)
 
 
 def create_audit_directory():
     """
     Creates directory for analyze_iam_policy audit files and places audit files there.
+
+    Essentially:
+    mkdir -p ~/.policy_sentry/audit
+    cp -r $MODULE_DIR/policy_sentry/shared/data/audit/ ~/.policy_sentry/audit/
     """
-    audit_directory_path = HOME + CONFIG_DIRECTORY + AUDIT_DIRECTORY_FOLDER
-    create_directory_if_it_doesnt_exist(audit_directory_path)
-    destination = audit_directory_path
+    create_directory_if_it_doesnt_exist(AUDIT_DIRECTORY_PATH)
+    destination = AUDIT_DIRECTORY_PATH
 
     existing_audit_files_directory = os.path.abspath(os.path.dirname(__file__)) + '/data/audit/'
     source = existing_audit_files_directory
@@ -51,6 +73,9 @@ def create_audit_directory():
 def create_policy_analysis_directory():
     """
     Creates directory for analyze_iam_policy policies.
+
+    Essentially:
+    mkdir -p ~/.policy_sentry/analysis
     """
     policy_analysis_directory_path = HOME + CONFIG_DIRECTORY + 'analysis'
     if os.path.exists(policy_analysis_directory_path):
@@ -62,6 +87,9 @@ def create_policy_analysis_directory():
 def create_default_overrides_file():
     """
     Copies over the overrides file in the config directory
+
+    Essentially:
+    cp $MODULE_DIR/policy_sentry/shared/data/access-level-overrides.yml ~/policy_sentry/access-level-overrides.yml
     """
     existing_overrides_file_name = 'access-level-overrides.yml'
     target_overrides_file_path = HOME + CONFIG_DIRECTORY + existing_overrides_file_name
@@ -74,6 +102,9 @@ def create_default_overrides_file():
 def create_default_report_config_file():
     """
     Copies over the default report config file to the config directory
+
+    Essentially:
+    cp $MODULE_DIR/policy_sentry/shared/data/audit/report-config.yml ~/policy_sentry/audit/report-config.yml
     """
     existing_report_config_file = 'report-config.yml'
     target_report_config_file_path = AUDIT_DIRECTORY_PATH + existing_report_config_file
@@ -151,6 +182,7 @@ def override_access_level(
     """
     Given the service-specific override config, determine whether or not the
     override config tells us to override the access level in the documentation.
+
     :param service_override_config: Given that the name
     :param action_name: The name of the action
     :param provided_access_level: Read, Write, List, Tagging, or 'Permissions management'.